fix: test utils should not be bundled in prod build
