+++ frontend/src/tests/lib/components/canister-detail/RemoveCanisterControllerButton.spec.ts
@@ -5,7 +5,7 @@
 import { removeController } from "$lib/services/canisters.services";
 import { fireEvent } from "@testing-library/dom";
 import { render, waitFor } from "@testing-library/svelte";
-import { clickByTestId } from "../../testHelpers/clickByTestId";
+import { clickByTestId } from "../../../mocks/utils.mock";
 import RemoveCanisterControllerButton from "./RemoveCanisterControllerButtonTest.svelte";
 
 jest.mock("$lib/services/canisters.services", () => {
