+++ frontend/src/tests/lib/components/project-detail/ParticipateButton.spec.ts
@@ -16,7 +16,7 @@ import {
   summaryForLifecycle,
 } from "../../../mocks/sns-projects.mock";
 import { renderContextCmp } from "../../../mocks/sns.mock";
-import { clickByTestId } from "../../testHelpers/clickByTestId";
+import { clickByTestId } from "../../../mocks/utils.mock";
 
 describe("ParticipateButton", () => {
   it("should render project participate button", () => {
