+++ frontend/src/tests/lib/modals/canisters/AddCyclesModal.spec.ts
@@ -16,7 +16,7 @@ import {
   mockSubAccount,
 } from "../../../mocks/accounts.store.mock";
 import { renderModal } from "../../../mocks/modal.mock";
-import { clickByTestId } from "../../testHelpers/clickByTestId";
+import { clickByTestId } from "../../../mocks/utils.mock";
 import AddCyclesModalTest from "./AddCyclesModalTest.svelte";
 
 jest.mock("$lib/services/canisters.services", () => {
