--- frontend/src/tests/lib/testHelpers/clickByTestId.ts
@@ -1,11 +1,0 @@
-import { fireEvent } from "@testing-library/dom";
-
-export const clickByTestId = async (
-  queryByTestId: (matcher: string) => HTMLElement | null,
-  testId: string
-) => {
-  const element = queryByTestId(testId);
-  expect(element).toBeInTheDocument();
-
-  element && fireEvent.click(element);
-};
